+++ ameriflux_pipeline/eddypro/eddyproformat.py
@@ -188,6 +188,7 @@ class EddyProFormat:
         Returns:
             df (object): Pandas DataFrame object
         """
+        df = df.astype({'TIMESTAMP': str})
         df['TIMESTAMP'] = df['TIMESTAMP'].map(lambda t: t.replace('/', '-'))
         df['TIMESTAMP'][0] = 'yyyy-mm-dd HH:MM'  # Change unit TS to yyyy-mm-dd HH:MM to match eddypro format
         return df