+++ ameriflux_pipeline/master_met/mastermetprocessor.py
@@ -7,6 +7,10 @@
 import pandas as pd
 import numpy as np
 from datetime import timedelta
+from pandas.api.types import is_datetime64_any_dtype as is_datetime64
+
+from utils.data_validation import DataValidation
+import utils.data_util as data_util
 
 import warnings
 warnings.filterwarnings("ignore")
@@ -47,6 +51,9 @@ class MasterMetProcessor:
         # get meta data
         # NOTE 2
         df_meta, file_meta = MasterMetProcessor.get_meta_data(file_df_meta)
+        if df_meta is None:
+            print("Please check met data file. Aborting")
+            return None, None
         # NOTE 4
         df_meta = MasterMetProcessor.add_U_V_units(df_meta)
 
@@ -54,7 +61,8 @@ class MasterMetProcessor:
         user_confirmation = user_confirmation.lower()
         df_precip = MasterMetProcessor.read_precip_data(input_precip_path, precip_lower, precip_upper,
                                                         missing_time_threshold, user_confirmation)
-
+        if df_precip is None:
+            print("Merging of precipitation data is not possible.")
         # change column data types
         df = MasterMetProcessor.change_datatype(df)
 
@@ -84,61 +92,80 @@ class MasterMetProcessor:
         # step 5 in guide. Calculation of soil heat flux
         # TODO : test with old data (non-critical)
         if MasterMetProcessor.soil_heat_flux_check(df):
-            df['shf_1_Avg'] = MasterMetProcessor.soil_heat_flux_calculation(df['shf_mV_Avg(1)'], df['shf_cal_Avg(1)'])
+            try:
+                shf_mV, shf_cal = df['shf_mV_Avg(1)'], df['shf_cal_Avg(1)']
+                df['shf_1_Avg'] = MasterMetProcessor.soil_heat_flux_calculation(shf_mV, shf_cal)
+            except KeyError:
+                print("Soil heat flux calculation failed. Check if columns shf_mV_Avg(1) and shf_cal_Avg(1) exists.")
 
         # Step 6 in guide. Absolute humidity check
-        df['Ah_fromRH'] = MasterMetProcessor.AhFromRH(df['AirTC_Avg'], df['RH_Avg'])
-        # add Ah_fromRH column and unit to df_meta
-        Ah_fromRH_unit = 'g/m^3'
-        df_meta['Ah_fromRH'] = Ah_fromRH_unit
+        try:
+            T, RH = df['AirTC_Avg'], df['RH_Avg']
+            df['Ah_fromRH'] = MasterMetProcessor.AhFromRH(T, RH)
+            # add Ah_fromRH column and unit to df_meta
+            Ah_fromRH_unit = 'g/m^3'
+            df_meta['Ah_fromRH'] = Ah_fromRH_unit
+        except KeyError:
+            print("AhFromRH calculation failed. Check if columns 'AirTC_Avg' and 'RH_Avg' exists.")
 
         # Step 4 in guide
         df = MasterMetProcessor.replace_empty(df)
 
         # NOTE 6
-        # delete newly created temp variables
+        # delete temporarily created variables
         df = MasterMetProcessor.delete_new_variables(df, new_variables)
 
-        # step 8 in guide - add precip data. join df and df_precip
-        # keep all met data and have NaN for precip values that are missing - left join with met data
-        # throw a warning if there are extra timestamps in met data
-        if (df.shape[0] > df_precip.shape[0]):
-            # there are more records in met data
-            print("Extra timestamps in met data. Joining precip with NaN value in extra timestamps")
-        # NOTE 8
-        df = pd.merge(df, df_precip, on='TIMESTAMP', how='left')
-
-        # add precipitation unit mm to df_meta
-        df_meta['Precip_IWS'] = 'mm'
+        if df_precip is not None:
+            # step 8 in guide - add precip data. join df and df_precip
+            # keep all met data and have NaN for precip values that are missing - left join with met data
+            # throw a warning if there are extra timestamps in met data
+            if (df.shape[0] > df_precip.shape[0]):
+                # there are more records in met data
+                print("Extra timestamps in met data. Joining precip with NaN value in extra timestamps")
+            # NOTE 8
+            df = pd.merge(df, df_precip, on='TIMESTAMP', how='left')
+            # add precipitation unit mm to df_meta
+            df_meta['Precip_IWS'] = 'mm'
 
         # step 7 in guide - calculation of shortwave radiation
         SW_unit = 'W/m^2'  # unit for shortwave radiation
         # NOTE 10
         if 'SWUp_Avg' in df.columns:
             shortwave_out = 'SWUp_Avg'
+            df['SW_out_Avg'] = df[shortwave_out]
+            df_meta['SW_out_Avg'] = SW_unit  # add shortwave radiation units
         elif 'CM3Dn_Avg' in df.columns:
             shortwave_out = 'CM3Dn_Avg'
-        df['SW_out_Avg'] = df[shortwave_out]
-        df_meta['SW_out_Avg'] = SW_unit  # add shortwave radiation units
-        albedo_col = str(df.filter(regex=("Albedo|albedo|ALB")).columns[0])
-        if albedo_col not in df.columns:
+            df['SW_out_Avg'] = df[shortwave_out]
+            df_meta['SW_out_Avg'] = SW_unit  # add shortwave radiation units
+        albedo_col = df.filter(regex="Albedo|albedo|ALB").columns.to_list()
+        if not albedo_col or albedo_col[0] not in df.columns:
             # calculate albedo_avg from shortwave out and shortwave in
             if 'SWDn_Avg' in df.columns:
                 shortwave_in = 'SWDn_Avg'
             elif 'CM3Up_Avg' in df.columns:
                 shortwave_in = 'CM3Up_Avg'
-            # avoid zero division error
-            df[albedo_col] = df.apply(lambda x: float(x[shortwave_out]) / float(x[shortwave_in])
-                                      if float(x[shortwave_in]) != 0 else np.nan, axis=1)
+        try:
+            if shortwave_in and shortwave_out:
+                # avoid zero division error
+                df[albedo_col] = df.apply(lambda x: float(x[shortwave_out]) / float(x[shortwave_in])
+                                          if float(x[shortwave_in]) != 0 else np.nan, axis=1)
 
-        df_meta[albedo_col] = SW_unit  # add shortwave radiation units
+                df_meta[albedo_col] = SW_unit  # add shortwave radiation units
+        except NameError:
+            if albedo_col[0] not in df.columns:
+                print("Albedo present in met data. Shortwave not calculated")
+            else:
+                print("Shortwave calculation failed. Check SW or CD3 columns in met data.")
 
         # NOTE 2
         # concat the meta df and df if number of columns is the same
         if df_meta.shape[1] == df.shape[1]:
             df = pd.concat([df_meta, df], ignore_index=True)
         else:
-            print("Meta and data file columns not matching")
+            print("Number of columns in met data {} not the same as number of columns in meta data {}".
+                  format(df.shape[1], df_meta.shape[1]))
+            return None, None
 
         # return processed and merged df. should contain 81 columns
         return df, file_meta
@@ -158,7 +185,8 @@ class MasterMetProcessor:
         df = pd.read_csv(data_path, header=None)  # read file without headers.
 
         # process df to get meta data
-        file_df_meta = df.head(3)
+        # TODO : Check with Bethany if Min/Avg is to be checked for.
+        file_df_meta = df.head(4)  # first four lines of file contains meta data
         # the first row contains the meta data of file. second and third row contains met variables and their units
         file_df_meta.fillna('', inplace=True)  # fill NaNs with empty string for ease of replace
         file_df_meta = file_df_meta.applymap(lambda x: str(x).replace('"', ''))  # strip off quotes from all values
@@ -185,14 +213,21 @@ class MasterMetProcessor:
             df_meta (obj) : meta data of met data. Consists of column names and units.
             file_meta (obj) : meta data of file. Consists of file name, field site, and crop.
         """
+        # TODO : Check meta_data format. At present the code checks for first 4 lines.
+        # But even if the last line is not present, it is ok.
+        # Need to check at which row the timestamp/ numerical data is starting.
         file_meta = file_df_meta.head(1)
         # the first row contains meta data of file. Used to match the filename to soil key.
         # returned with the processed df
         df_meta = file_df_meta.iloc[1:, :]
+        if not DataValidation.is_valid_meta_data(df_meta):
+            print("Meta data not in valid format")
+            return None, None
         # second and third row contains meta data of met tower variables (column names and units)
         df_meta.columns = df_meta.iloc[0]
         df_meta.drop(df_meta.index[0], inplace=True)
         df_meta.reset_index(drop=True, inplace=True)  # reset index after dropping first row
+        df_meta = df_meta.head(1)  # dropping the last row of Min / Avg
         return df_meta, file_meta
 
     @staticmethod
@@ -222,27 +257,28 @@ class MasterMetProcessor:
             data_path (str): input data file path
             precip_lower (int) : Lower threshold value for precipitation in inches
             precip_upper (int) : Upper threshold value for precipitation in inches
-            missing_timeslot_threshold (int): Value for missing timeslot threshold. used for insert_missing_time method
+            missing_time_threshold (int): Value for missing timeslot threshold. used for insert_missing_time method
             user_confirmation (str) : Option to either insert or ignore missing timestamps
         Returns:
             obj: Pandas DataFrame object
         """
         df = pd.read_excel(data_path)  # read excel file
-        station_col = str(df.filter(regex=("Station|station")).columns[0])
-        df.drop([station_col], axis=1, inplace=True)  # drop unwanted columns
+        df = MasterMetProcessor.get_valid_precip_data(df)
+        if df is None:
+            print("Precipitation data not valid.")
+            return None
+
         # TODO: Ask Bethany - if missing time threshold for precip data is ok to be same as met data
         # NOTE 5
         # perform qa qc checks for precip data
         df = MasterMetProcessor.precip_qaqc(df, precip_lower, precip_upper, missing_time_threshold, user_confirmation)
         # convert precipitation from in to mm
         # TODO : import cf_units and use to convert units. / udunits
-        precip_col = str(df.filter(regex=("Precipitation|precipitation")).columns[0])
-        df['Precipitation (mm)'] = df[precip_col] * 25.4  # convert inches to millimeter
-        df.drop([precip_col], axis=1, inplace=True)  # drop unwanted columns
+        df['Precipitation_mm'] = df['Precipitation_in'] * 25.4  # convert inches to millimeter
+        df.drop(['Precipitation_in'], axis=1, inplace=True)  # drop unwanted columns
         # convert 5min samples to 30min samples by taking the sum
-        time_col = str(df.filter(regex=("Time|time|CST")).columns[0])
-        df = df.set_index(time_col)
-        precip_series = pd.Series(df['Precipitation (mm)'], index=df.index)
+        df = df.set_index('Timestamp')
+        precip_series = pd.Series(df['Precipitation_mm'], index=df.index)
         # resampling to 30min timeslots. 00-30 is summed and stored in 00min. (beginning of timestamp)
         # skipna False accounts for NaN in values. If NaN present, the 30min resample has value of NaN.
         precip_30 = precip_series.resample('30min').agg(pd.Series.sum, skipna=False)
@@ -253,6 +289,67 @@ class MasterMetProcessor:
         # replace / with - to match timestamp format of met data
         df['TIMESTAMP'] = df['TIMESTAMP'].map(lambda t: t.replace('-', '/'))
         return df
+
+    @staticmethod
+    def get_valid_precip_data(df):
+        """
+        Method to check if the input dataframe containing precipitation data is in valid format.
+        Checks for expected columns like Time and Precip columns, and expected datatypes for the columns
+        Returns the processed df
+        Args:
+            df (obj): Pandas dataframe object to check for valid format
+        Returns:
+            df (object): Pandas dataframe object which is the valid precip data with required columns.
+        """
+        time_flag, precip_flag = False, False
+        # check for timestamp and precip column
+        precip_col = df.filter(regex='Precipitation|precipitation|Precip|precip|Rain|rain|IWS').columns.to_list()
+        time_col = df.filter(regex='Date|Time|time|CST|timestamp|TIMESTAMP|Timestamp').columns.to_list()
+        if not time_col:
+            print("Timestamp column not present in Precipitation data.")
+            return None
+        if not precip_col:
+            print("Precipitation column not present in Precipitation data.")
+            return None
+
+        # there are more than 1 column that matches timestamp.
+        # Process the first column that matches the criteria and break
+        for col in time_col:
+            if is_datetime64(df[col]):
+                # col is of type datetime
+                df['Timestamp'] = df[col]
+                time_flag = True
+                break
+            elif DataValidation.string_validation(df[col].iloc[df[col].first_valid_index()]):
+                # parse only accepts str input. Check if the column type is string.
+                df['Timestamp'] = df[col].apply(lambda x: data_util.get_valid_datetime(x))
+                time_flag = True
+                break
+
+        # there are more than 1 column that matches precipitation. There could be precip in inches and mm.
+        for col in precip_col:
+            if DataValidation.float_validation(df[col].iloc[df[col].first_valid_index()]):
+                if any(inch_unit in col for inch_unit in ['(in)', 'inches', '(inches)']):
+                    df['Precipitation_in'] = df[col]
+                    precip_flag = True
+                    break
+                elif any(mm_unit in col for mm_unit in ['(mm)', 'mm', 'millimeter', 'millimeters',
+                                                        '(millimeter)', '(millimeters)']):
+                    df['Precipitation_in'] = df[col] / 25.4  # convert mm to inches
+                    precip_flag = True
+                    break
+
+        # all validations done
+        if time_flag and precip_flag:
+            return df[['Timestamp', 'Precipitation_in']]
+        elif not time_flag:
+            print("Precipitation timestamp not in correct format")
+            return None
+        elif not precip_flag:
+            print("Precipitation values not in correct format")
+            return None
+        else:
+            return None
 
     @staticmethod
     def precip_qaqc(df, precip_lower, precip_upper, missing_time_threshold, user_confirmation):
@@ -266,16 +363,15 @@ class MasterMetProcessor:
             df (obj): input precip dataframe
             precip_lower (float) : Lower threshold value for precipitation in inches
             precip_upper (float) : Upper threshold value for precipitation in inches
-            missing_timeslot_threshold (int): Value for missing timeslot threshold. used for insert_missing_time method
+            missing_time_threshold (int): Value for missing timeslot threshold. used for insert_missing_time method
             user_confirmation (str) : Option to either insert or ignore missing timestamps
         Returns:
             obj (Pandas DataFrame object): processed and cleaned precip dataframe
         """
         # check timestamps, if present for every 5 min
-        time_col = str(df.filter(regex=("Time|time|CST")).columns[0])
-        df['timedelta'] = MasterMetProcessor.get_timedelta(df[time_col])
+        df['timedelta'] = MasterMetProcessor.get_timedelta(df['Timestamp'])
         df, insert_flag = \
-            MasterMetProcessor.insert_missing_timestamp(df, time_col, 5.0,
+            MasterMetProcessor.insert_missing_timestamp(df, 'Timestamp', 5.0,
                                                         missing_time_threshold, user_confirmation)
         if insert_flag == 'N':
             # user confirmed not to insert missing timestamps.
@@ -284,12 +380,11 @@ class MasterMetProcessor:
         df.drop(['timedelta'], axis=1, inplace=True)
         # check precip values in between 0 and 0.2 in
         # get indexes where precip is greater than 0.2 or less than 0.
-        precip_col = str(df.filter(regex=("Precipitation|precipitation")).columns[0])
-        invalid_indexes = df.index[(df[precip_col] > precip_upper)].to_list()
-        invalid_indexes.extend(df.index[(df[precip_col] < precip_lower)].to_list())
+        invalid_indexes = df.index[(df['Precipitation_in'] > precip_upper)].to_list()
+        invalid_indexes.extend(df.index[(df['Precipitation_in'] < precip_lower)].to_list())
         # replace precip value with NaN at invalid indexes
         for index in invalid_indexes:
-            df[precip_col].iloc[index] = np.nan
+            df['Precipitation_in'].iloc[index] = np.nan
         # return cleaned df
         return df