+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added reading xls formating functionality [#164](https://github.com/ncsa/ameriflux-pipeline/issues/164)
 - Added details on met_data_merge in README [#178](https://github.com/ncsa/ameriflux-pipeline/issues/178)
 - Data validation for met data merge [#168](https://github.com/ncsa/ameriflux-pipeline/issues/168)
+- Data validation for creation of master met data [#172](https://github.com/ncsa/ameriflux-pipeline/issues/172)
 
 ### Fixed
 - Python packaging of the codes so it can run in command prompt. [#70](https://github.com/ncsa/ameriflux-pipeline/issues/70)